Include annotations in function documentation
This would need to be tweaked to looks better: the annotations are represented using the built‑in `repr`, which is not always good at it (ex. when the annotation is a type, like `int`, it displays as `<class 'int'>` instead of just as `int`). There is no way to workaround this using a custom `__repr__` or `__str__` of a custom class, as the annotations are ordinary tuples. May be an hook formater function could be an option. I don't know if there is anything like plug-ins with *pdoc*; if there are, this may be an option.