Fix wrong HTML body height
The `height: 100%;` is erroneous, as it gives the body an height which is 100% of that of the view-port, which is not relevant for a content which is to be scrollable. To see it, one may give the body a non-white background color, say orange (as an example), then scroll the content: the top has an orange and the rest of the scrollable content has a white (default) background.