Revert "Switch to Markdown2"
This reverts commit 6b0b0c3d33be3816c05a59f5dcb9d3cda3c2b6e9.

Compared to Python-Markdown, which is [maintained],
Markdown2 is [bug-ridden]. Incorrectly renders at least:

    [`code`](links)

which is important for us. The owner seems to [think it] an
edge case despite numerous such reports.

Python-Markdown has [12k dependents] whereas Markdown2 has
[only 1.3k], making the former's community quite a bit larger.

Finally, the [initial reason] for switch to Markdown2 was its
handling of fenced code blocks (with an addon). But
Python-Markdown supports this case just the same, with an addon.

[maintained]: https://github.com/Python-Markdown/markdown/issues
[bug-ridden]: https://github.com/trentm/python-markdown2/issues
[think it]: https://github.com/trentm/python-markdown2/issues/259
[12k dependents]: https://libraries.io/pypi/Markdown/usage
[only 1.3k]: https://libraries.io/pypi/markdown2/usage
[initial reason]: https://github.com/mitmproxy/pdoc/pull/119