Fix blank lines breaking indented sections within fenced code blocks
The issue was if your code was indented in any way (eg: inside an if statement) and surrounded by blank lines,
it would be detected as another code block. This was due to the regular expression used to process fenced code blocks not
detecting the "<code></code>" tags surrounding that section.

See https://github.com/mitmproxy/pdoc/pull/267 for more details.

The regex lookahead in `pdoc.markdown2.Markdown._do_code_blocks` has been tweaked like so: `(?![^<]*\</code\>)` -> `(?![^*]*\</code\>)`
Previously, it would (I think) check the next HTML tag in the string to see if it is a `</code>` tag.
The new lookahead will check the entire string for ANY `</code>` tags.