Suggested amends regarding the "pdoc3" controversy
This aims to provide additional context to the situation regarding pydoc3 - whilst being more sensitive to the cultural significance of the symbolism that has sadly been conflated with something quite different. It's with some irony that the existing text states that "everyone is treated with respect"-  but there's been a lack of meaningful dialogue over quite a sensitive topic. There's a full discussion in issue #330.