Module.members: Suppress underscored members
The [documentation](https://pdoc.dev/docs/pdoc.html#control-what-is-documented)
states:

> If __all__ is not defined, then pdoc will consider all members public that do
not start with an underscore and are defined in the current module (i.e. they
are not imported).

Existing behavior is inconsistent; classes in modules prefixed with
underscore, for example, are exposed as members of that class; see attached
falsifying test case.

I've proposed a fix here in the `pdoc.doc.Module._member_objects`
implementation. Maintainers, please let me know if there's anything else to fix
up.