feat: include documentation for double-underscore methods when requested
1.  Add `--include-dunder` Boolean option to command line args to request
    that double-underscore (dunder) methods with docstrings be included.

2.  Store value as `include_dunder` in environment passed to Jinja2.

3.  Modify Jinja2 template to include dunder methods if option enabled.

4.  Add test for new feature.

This does _not_ work as intended at the moment because inherited
dunder methods with docstrings are displayed as well as those having
dunder methods that are defined in the class. As a result, methods
like `__dict__` and `__weakref__` show up for a class that only
defines a docstring for `__lt__` (and the docs for `__eq__` also show
up, even though the test explicitly does *not* have a docstring).
Guidance would be greatly appreciated.

Closes #654